+++ ICD.Connect.Routing.Extron/Devices/Switchers/SwUsb/AbstractSwUsbSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.SwUsb
+{
+	public abstract class AbstractSwUsbSettings : AbstractExtronSwitcherDeviceSettings, ISwUsbSettings
+	{
+	}
+}