+++ ICD.Connect.Routing.Extron/Devices/Switchers/SwUsb/ISwUsbSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.SwUsb
+{
+	public interface ISwUsbSettings : IExtronSwitcherDeviceSettings
+	{
+	}
+}