+++ ICD.Connect.Routing.Extron/Devices/Switchers/SwUsb/Sw2Usb/Sw2UsbSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.Extron.Devices.Switchers.SwUsb.Sw2Usb
+{
+	[KrangSettings("SW2USB", typeof(Sw2UsbDevice))]
+	public sealed class Sw2UsbSettings : AbstractSwUsbSettings
+	{
+	}
+}