+++ ICD.Connect.Routing.Extron/Devices/Switchers/SwUsb/Sw4Usb/Sw4UsbSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.Extron.Devices.Switchers.SwUsb.Sw4Usb
+{
+	[KrangSettings("SW4USB", typeof(Sw4UsbDevice))]
+	public sealed class Sw4UsbSettings : AbstractSwUsbSettings
+	{
+	}
+}